This is a partial fix for github issue 11288
It appears to resolve the basic issue where you have NGEN image A that inlined
code from non-NGEN'ed IL image B. However I am concerned there are more
complicated cases that are not yet covered. Module::GetAssemblyIfLoaded has a
second section of code that handles versioning mismatches and the profiler path
deliberately opts in to enabling that fallback section. If we had a case that
needed the fallback I think it would hit the same load-level issue that the
primary code path hit.