Convert ListLock to use templated data
The current ListLock only supports locking on pointers. As a pre-requisite to making the code generation pipeline operate on code versions rather than just MethodDesc* we need the ListLock used in that path to be able to take a lock on the NativeCodeVersion structure. Templating the ListLock creates the flexibility we will need.