Eliminate a bunch of interpreter special casing throughout the MethodDesc code in preparation for refactoring there.
Instead of the customized method switching protocol the interpreter was previously using I am standardizing it as part of the tiered compilation flow. It is still my intention that tiered compilation doesn't use interpreter by default, but it will work when FEATURE_TIERED_COMPILATION and FEATURE_INTERPRETER are both enabled.