Bind the rejit no-inlining flag properly
ReJIT has long had a race condition where it assumes that requests to rejit and then the JIT compilation of those requests will be temporally ordered. Instead of packaging all the configuration into the flags for JitUnsafeFunction it has CEEJitInfo call back to rejit manager to get the current setting of a no-inline flag. If a profiler called RequestReJIT while the JIT was in progress it could easily update the state of the no inlining flag and get the in-progress JIT execution to use the new state rather than the state which was originally specified for that code body. If left unfixed Tiered Compilation would exacerbate the bug because it could be jitting the Tier1 version of the default IL, pick up the no-inlining flag from a non-default IL version, then when the profiler reverts the process will be running code with the wrong jit flags arbitrarily long into the future.