Remove the explicit reverted state from ILCodeVersion rejit state
The ILCodeVersionManager already has a concept of the active IL code version (CodeVersionManager::GetActiveILCodeVersion) that I'd like to keep seperate from the particular configuration flags used in a given stage. This also simplifies code because if we can identify the active version with a direct pointer deref (not yet implemented) rather than needing to loop over all versions. If there were a large number of IL versions we also move from O(N) to O(1) lookup.