Implement most of the CodeVersionManager object model
Flesh out the implementation of the various version objects, the tree nodes that back them + the collections and iterators. Instead of the ReJitManager technique that adds N versions of the same method to an shash and then iterates the hashtable bucket, this approach hashes a single MethodDescVersioningState or ILCodeVersioningState per-method, then chains a linked list of versions from there.