Bug fixes
Contract violation in ReJitManager::BindILVersion because GetIL() throws. Added GetILNoThrow().

Contract violation in CodeVersionManager::EnumerateClosedMethodDescs, it is OK for the pMD to be NULL.

CodeVersionManager::EnumerateClosedMethodDescs() was incorrectly filtering out MethodDescs that didn't have native code. This was legitimate when all methods versioned with jump stamps because it is impossible to jumpstamp an instantiation that doesn't have the default code generated. However with FixupPrecode versioning it is possible to immediately patch a method to a non-default version such that HasNativeCode remains false.