WIP
This moves JitBench away from using CreateStore, and instead downloads AspNetCore directly from Azure, already R2R compiled. The only remaining issue is AllReady, which is currently trying to restore asp.net 2.0 as part of its build. We should probably patch it so that similar to MusicStore it will honor whatever version of ASP.Net is set in the JITBENCH_ASPNET_VERSION environment variable.

In the short term its a little annoying because now we are tracking an ASP.Net version too in dependencies.props, but I think as soon as we hit 2.1 we can just pin to the release version and leave it there.