Refactoring/Renaming to better promote that versioning isn't only for tiering even if that is the current use-case. In the future I'd advocate that we collapse EnC, ReJIT, and tiering to all use the same versioning primitives. Anything that couples method versionability to tiering specifically makes it harder to move in that direction.
IsTieredMethodVersionableWithPrecode -> IsVersionableWithPrecode
  - Now callable for any method, not only TieredCompilation eligible methods, and TRUE shouldn't imply the method is eligible for tiering

IsTieredMethodVersionableWithVtableSlotBackpatch -> IsVersionableWithCallerSlots
  - Now callable for any method, not only TieredCompilation eligible methods, and TRUE shouldn't imply the method is eligible for tiering
  - Also eliminated the implied coupling to vtable slots since these methods use vtable, VSD, and FuncPtrStub slots

IsTieredVTableMethod -> IsVersionableWithCallerSlots
  - TRUE doesn't imply that the method is eligible for tiering

ResetTieredMethodCodeEntryPoint -> ResetMethodCodeEntrypoint
   - Function should work for any method that uses Precode or CallerSlot versioning

As best I could tell, none of the sites that were using IsTiered* methods truly required the methods to be used in tiering, they only required certain versioning properties to hold so the switch was relatively straightforward. In some cases I switched to using GetVersioningTechnique() to more efficiently test against several options but calling multiple IsVersionableWith* methods would also work.  Performance-wise I think the new code generally evaluates the same or fewer conditionals than it did before, not counting code inside ASSERTE statements and I didn't check if it impacted the compiler's inlining quality. The switch from IsEligibleForTieredCompilation() to pMethod->IsVersionableWithPrecode() || pMethod->IsVersionableWithCallerSlots() in codeversion.cpp might be a hard one for the compiler to reduce automatically, but I suspect the path is cold enough that it wouldn't matter or there are alternate ways to write it that could make the inlining a little more explicit to the compiler.

I made a few other small changes while I was in there looking at the code. Mostly it was comments but I did remove an unneeded call to IsEnCMethod() in DetermineAndSetIsEligibleForTieredCompilation(), and removed the dependency in IsNativeCodeStableAfterInit on IsEligibleForTieredCompilation().