Updating NetPerf format (Step 3)
Add support for 64 bit thread ids that are used on Mac. I don't have a mac handy so this hasn't been tested yet, but in the debugger I can confirm that 64 bits is being allocated in the serialized format. I tried to use a relatively minimalist approach that propagates the 64 bit ID only through the code paths that were required to reach EventPipe. As soon as code branches off the EventPipe path I truncated the ID to 32 bits to preserve prior behavior. Other areas of the runtime might benefit from having the 64 bit ID available but I didn't include that in my scope of work.