Updating NetPerf format (Step 4)
In order to stream for long periods of time we need to ensure that the TraceEvent reader process has constant memory usage regardless of the number of events. Previously every EventBlock and SerializationType that preceeded it were added to an interning table. When doing reasonably low latency event transmission we might emit ~100 blocks/sec leading to a leak of ~864MB/day. This commit fixes that by marking the objects private so they will not be interned. The format already ensures there are no intra-stream object references so the intern table was not serving any purpose.