Updating NetPerf format (Step 7)
In preparation for adding sequence numbers it was helpful to refactor EventPipeThread a bit. I wanted to add a new per-thread per-session field (sequence number) and the current pattern involves creating a new dictionary for each field. In addition to the general overhead of maintaining multiple parallel dictionaries it isn't an easy structure to access lock-lessly and I wanted to be able to do cross-thread reads of the sequence numbers. To remedy that and just do a little general house cleaning I moved things around like this:
a) EventPipeThread moved to its own file
b) All the per-session state for a thread moved to EventPipeThreadSessionState. Each thread has an array of 64 pointers to session states, one for each possible session.
c) EventPipeBufferManager now tracks a list of EventPipeThreadSessionState instead of EventPipeBufferList.
d) Added GetIndex() on session to distinguish it from GetId(), GetIndex() returns a number 0-63