Updating NetPerf format (Step 8)
Add a sequence number counter to a thread. Each emitted event gets assigned a number which flows with it through the buffer manager and is emitted into the file format. If the event is dropped because a buffer is not allocated then the number will be missing from the stream. We also attach a thread list with current sequence numbers to the sequence point blocks so that the reader can identify if a thread is dropping events regardless of whether any events are getting through.